Implementation of Robin Hood hashing for AWS C Common (#17)
* Improve test macro hygeine

* Add forward declarations to test declaration macros

* Add saturating/checked multiply helpers

* Add implementation of robin hood hash tables

* Fixed windows build for the hash table implementation.

* We shouldn't build header files.

Without this, we endup with unused function errors because we attempt to
compile .c files that are meant for inclusion within other .c files.

* fix gcc warnings

* fix assembly to work on gcc

* needlessly initialize variables to shutup gcc silly warnings.

* conditionally enable HW optimization with inline assembly

* variable maybe uninitialized

* Avoid double usage of rax/rdx for saturation value

* replace C++ style comments with C

* use jnc instead of cmov instruction for 32-bit

cmov is not always available on 32bit machines, and also benchmarking
cmove against jnc yields no significant difference; in fact,
in some cases (clang compiles) it resulted in faster code.

Removing the #ifdef for cmov allows us to test the 32bit code on 64bit
machines used by developers and CI. Which avoids surprise errors like
the compile error of "label redefinition" which I also fixed in this commit.

* make cppcheck happy

* use early clobbers & change C++ style comments

* Additional improvements to saturated/checked multiplication routines

* Use symbolic names in some places
* Fix more missing earlyclobber operands

* Hash table API consistency improvements

* Use aws_raise_error to report errors and update tests accordingly)
* Use bitmask for foreach callback, and add a test for it
* Code style changes

* More macro hygiene fixes

* Indicate item-not-present through means other than raising errors

* Fix compiler warning in tests

* Fix cppcheck errors

* make cppcheck happy

* Constrain some operands to registers

* make cppcheck happy again

* supress memory leak flag by cppcheck

* whitespace

+++ tests/priority_queue_test.c
@@ -103,7 +103,7 @@ static int test_priority_queue_random_values(struct aws_allocator *alloc, void *
     int storage[SIZE], err;
     aws_priority_queue_static_init(&queue, storage, SIZE, sizeof(int), compare_ints);
     int values[SIZE];
-    srand((size_t)&queue);
+    srand((unsigned)(uintptr_t)&queue);
     for(int i = 0; i < SIZE; i++) {
         values[i] = rand() % 1000;
         err = aws_priority_queue_push(&queue, &values[i]);
@@ -146,7 +146,7 @@ static int test_priority_queue_size_and_capacity(struct aws_allocator *alloc, vo
     struct aws_priority_queue queue;
     int err = aws_priority_queue_dynamic_init(&queue, alloc, 5, sizeof(int), compare_ints);
     ASSERT_SUCCESS(err, "Dynamic init failed with error %d", err);
-    int capacity = aws_priority_queue_capacity(&queue);
+    size_t capacity = aws_priority_queue_capacity(&queue);
     ASSERT_INT_EQUALS(5, capacity, "Expected Capacity %d but was %d", 5, capacity);
 
     for(int i = 0; i < 15; i++) {
@@ -154,7 +154,7 @@ static int test_priority_queue_size_and_capacity(struct aws_allocator *alloc, vo
         ASSERT_SUCCESS(err, "Push operation failed with error %d", err);
     }
 
-    int size = aws_priority_queue_size(&queue);
+    size_t size = aws_priority_queue_size(&queue);
     ASSERT_INT_EQUALS(15, size, "Expected Size %d but was %d", 15, capacity);
 
     capacity = aws_priority_queue_capacity(&queue);
